Refactored to add qanta module
The purpose of this module is to be the top level module for the qanta
codebase. It will serve the second purpose that any code in qanta will
be pylinted and when possible have unit tests run on it. That is to
say, qanta should represent code that is by some means vetted.

Additionally, as code gets moved to qanta it will be audited to check
if there are dead functions/methods/classes/code (nothing calls it). If
so, that code will be removed.